# API_blioteca-main/src/model/dto/UsuarioDto.ts
export class UsuarioDto{
    id: number;
    pessoaId: number;
    senha: string;

    constructor(id:number, pessoaId:number, senha:string){
        this.id = id;
        this.pessoaId = pessoaId;
        this.senha = senha;
    }
}